fix: move resolveAccountAddress feature to KeyringHandler scope only
