+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/snap-bitcoin-wallet.git"
   },
   "source": {
-    "shasum": "li+QpEut5Qn4fAGp0DkCFbGa6+RmarGlXAc7y37OuTU=",
+    "shasum": "W3irDqbovSMEOoZ/6MCfpwhu4lgSC8xwKnsAmsFoeGI=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
