+++ packages/snap/src/handlers/KeyringRequestHandler.test.ts
@@ -1,25 +1,29 @@
 import type { Txid, Psbt, Amount, LocalOutput } from '@metamask/bitcoindevkit';
-import type { KeyringRequest, KeyringAccount } from '@metamask/keyring-api';
-import { BtcMethod, BtcScope } from '@metamask/keyring-api';
+import type { KeyringRequest } from '@metamask/keyring-api';
 import { mock } from 'jest-mock-extended';
 import { assert } from 'superstruct';
 
 import type { AccountUseCases } from '../use-cases';
+import { KeyringRequestHandler } from './KeyringRequestHandler';
 import {
   BroadcastPsbtRequest,
   ComputeFeeRequest,
   FillPsbtRequest,
   GetUtxoRequest,
-  KeyringRequestHandler,
   SendTransferRequest,
   SignPsbtRequest,
-  type BtcWalletRequest,
-} from './KeyringRequestHandler';
+} from './validation';
 import type { BitcoinAccount, ConfirmationRepository } from '../entities';
 import { AccountCapability } from '../entities';
 import type { Utxo } from './mappings';
 import { mapToUtxo } from './mappings';
 import { parsePsbt } from './parsers';
+
+/* eslint-disable @typescript-eslint/naming-convention */
+jest.mock('@metamask/bitcoindevkit', () => ({
+  Address: { from_string: jest.fn() },
+  Amount: { from_btc: jest.fn() },
+}));
 
 jest.mock('superstruct', () => ({
   ...jest.requireActual('superstruct'),
@@ -182,7 +186,6 @@ describe('KeyringRequestHandler', () => {
     it('executes computeFee', async () => {
       mockAccountsUseCases.computeFee.mockResolvedValue(
         mock<Amount>({
-          // eslint-disable-next-line @typescript-eslint/naming-convention
           to_sat: () => BigInt(1000),
         }),
       );
@@ -544,132 +547,6 @@ describe('KeyringRequestHandler', () => {
         pending: false,
         result: { signature: 'signature' },
       });
-    });
-  });
-
-  describe('resolveAccountAddress', () => {
-    const mockAccount1 = mock<KeyringAccount>({
-      id: 'account-1',
-      address: 'test123',
-      scopes: [BtcScope.Regtest],
-    });
-    const mockAccount2 = mock<KeyringAccount>({
-      id: 'account-2',
-      address: 'test456',
-      scopes: [BtcScope.Regtest, BtcScope.Testnet],
-    });
-    const mockAccount3 = mock<KeyringAccount>({
-      id: 'account-3',
-      address: 'test789',
-      scopes: [BtcScope.Testnet],
-    });
-
-    it('resolves account address for SignPsbt', () => {
-      const request = {
-        method: BtcMethod.SignPsbt as const,
-        params: {
-          account: { address: 'test123' },
-          psbt: 'psbt',
-          options: { fill: false, broadcast: false },
-        },
-      };
-
-      const result = handler.resolveAccountAddress(
-        [mockAccount1, mockAccount2, mockAccount3],
-        BtcScope.Regtest,
-        request,
-      );
-
-      expect(result).toBe('bip122:regtest:test123');
-    });
-
-    it('resolves account address for SendTransfer', () => {
-      const request = {
-        method: BtcMethod.SendTransfer as const,
-        params: {
-          account: { address: 'test456' },
-          recipients: [{ address: 'recipient', amount: '1000' }],
-        },
-      };
-
-      const result = handler.resolveAccountAddress(
-        [mockAccount1, mockAccount2, mockAccount3],
-        BtcScope.Regtest,
-        request,
-      );
-
-      expect(result).toBe('bip122:regtest:test456');
-    });
-
-    it('resolves account address for SignMessage', () => {
-      const request = {
-        method: BtcMethod.SignMessage as const,
-        params: {
-          account: { address: 'test123' },
-          message: 'hello',
-        },
-      };
-
-      const result = handler.resolveAccountAddress(
-        [mockAccount1, mockAccount2, mockAccount3],
-        BtcScope.Regtest,
-        request,
-      );
-
-      expect(result).toBe('bip122:regtest:test123');
-    });
-
-    it('throws error if no accounts with scope', () => {
-      const request = {
-        method: BtcMethod.SignPsbt as const,
-        params: {
-          account: { address: 'test123' },
-          psbt: 'psbt',
-          options: { fill: false, broadcast: false },
-        },
-      };
-
-      expect(() =>
-        handler.resolveAccountAddress(
-          [mockAccount3],
-          BtcScope.Regtest,
-          request,
-        ),
-      ).toThrow('No accounts with this scope');
-    });
-
-    it('throws error if account address not found', () => {
-      const request = {
-        method: BtcMethod.SignPsbt as const,
-        params: {
-          account: { address: 'notfound' },
-          psbt: 'psbt',
-          options: { fill: false, broadcast: false },
-        },
-      };
-
-      expect(() =>
-        handler.resolveAccountAddress(
-          [mockAccount1, mockAccount2, mockAccount3],
-          BtcScope.Regtest,
-          request,
-        ),
-      ).toThrow('Account not found');
-    });
-
-    it('throws error for unsupported method', () => {
-      const request = {
-        method: 'unsupported' as BtcMethod,
-        params: {},
-      };
-
-      expect(() =>
-        handler.resolveAccountAddress(
-          [mockAccount1, mockAccount2, mockAccount3],
-          BtcScope.Regtest,
-          request as unknown as BtcWalletRequest,
-        ),
-      ).toThrow('Unsupported method');
     });
   });
 });
