+++ packages/snap/src/handlers/KeyringRequestHandler.ts
@@ -1,22 +1,6 @@
-import {
-  BtcMethod,
-  type KeyringAccount,
-  type KeyringRequest,
-  type KeyringResponse,
-} from '@metamask/keyring-api';
-import type { CaipAccountId, CaipChainId, Json } from '@metamask/snaps-sdk';
-import type { Infer } from 'superstruct';
-import {
-  array,
-  assert,
-  boolean,
-  literal,
-  number,
-  object,
-  optional,
-  string,
-  union,
-} from 'superstruct';
+import type { KeyringRequest, KeyringResponse } from '@metamask/keyring-api';
+import type { Json } from '@metamask/snaps-sdk';
+import { assert } from 'superstruct';
 
 import type { ConfirmationRepository } from '../entities';
 import {
@@ -26,138 +10,38 @@ import {
 } from '../entities';
 import { mapToUtxo } from './mappings';
 import { parsePsbt } from './parsers';
+import {
+  BroadcastPsbtRequest,
+  ComputeFeeRequest,
+  FillPsbtRequest,
+  GetUtxoRequest,
+  SendTransferRequest,
+  SignMessageRequest,
+  SignPsbtRequest,
+} from './validation';
 import type { AccountUseCases } from '../use-cases/AccountUseCases';
-
-/**
- * Wallet account struct for Bitcoin requests.
- */
-const WalletAccountStruct = object({
-  address: string(),
-});
-
-export const SignPsbtRequest = object({
-  account: WalletAccountStruct,
-  psbt: string(),
-  feeRate: optional(number()),
-  options: object({
-    fill: boolean(),
-    broadcast: boolean(),
-  }),
-});
 
 export type SignPsbtResponse = {
   psbt: string;
   txid: string | null;
 };
 
-export const ComputeFeeRequest = object({
-  account: WalletAccountStruct,
-  psbt: string(),
-  feeRate: optional(number()),
-});
-
 export type ComputeFeeResponse = {
   // Fee in satoshis
   fee: string;
 };
 
-export const BroadcastPsbtRequest = object({
-  account: WalletAccountStruct,
-  psbt: string(),
-});
-
 export type BroadcastPsbtResponse = {
   txid: string;
 };
-
-export const FillPsbtRequest = object({
-  account: WalletAccountStruct,
-  psbt: string(),
-  feeRate: optional(number()),
-});
 
 export type FillPsbtResponse = {
   psbt: string;
 };
 
-export const SendTransferRequest = object({
-  account: WalletAccountStruct,
-  recipients: array(
-    object({
-      address: string(),
-      amount: string(),
-    }),
-  ),
-  feeRate: optional(number()),
-});
-
-export const GetUtxoRequest = object({
-  account: WalletAccountStruct,
-  outpoint: string(),
-});
-
-export const SignMessageRequest = object({
-  account: WalletAccountStruct,
-  message: string(),
-});
-
 export type SignMessageResponse = {
   signature: string;
 };
-
-/**
- * Validates that a JsonRpcRequest is a valid Bitcoin request.
- *
- * TODO: update btc-methods.md to include all the new methods
- *
- * @see https://github.com/MetaMask/accounts/blob/main/packages/keyring-api/docs/btc-methods.md
- */
-export const SignPsbtKeyringRequestStruct = object({
-  method: literal(BtcMethod.SignPsbt),
-  params: SignPsbtRequest,
-});
-
-export const FillPsbtKeyringRequestStruct = object({
-  method: literal(BtcMethod.FillPsbt),
-  params: FillPsbtRequest,
-});
-
-export const ComputeFeeKeyringRequestStruct = object({
-  method: literal(BtcMethod.ComputeFee),
-  params: ComputeFeeRequest,
-});
-
-export const BroadcastPsbtKeyringRequestStruct = object({
-  method: literal(BtcMethod.BroadcastPsbt),
-  params: BroadcastPsbtRequest,
-});
-
-export const SendTransferKeyringRequestStruct = object({
-  method: literal(BtcMethod.SendTransfer),
-  params: SendTransferRequest,
-});
-
-export const GetUtxoKeyringRequestStruct = object({
-  method: literal(BtcMethod.GetUtxo),
-  params: GetUtxoRequest,
-});
-
-export const SignMessageKeyringRequestStruct = object({
-  method: literal(BtcMethod.SignMessage),
-  params: SignMessageRequest,
-});
-
-export const BtcWalletRequestStruct = union([
-  SignPsbtKeyringRequestStruct,
-  FillPsbtKeyringRequestStruct,
-  ComputeFeeKeyringRequestStruct,
-  BroadcastPsbtKeyringRequestStruct,
-  SendTransferKeyringRequestStruct,
-  GetUtxoKeyringRequestStruct,
-  SignMessageKeyringRequestStruct,
-]);
-
-export type BtcWalletRequest = Infer<typeof BtcWalletRequestStruct>;
 
 export class KeyringRequestHandler {
   readonly #accountsUseCases: AccountUseCases;
@@ -363,62 +247,5 @@ export class KeyringRequestHandler {
       pending: false,
       result,
     };
-  }
-
-  /**
-   * Resolves the address of an account from a signing request.
-   *
-   * This is required by the routing system of MetaMask to dispatch
-   * incoming non-EVM dapp signing requests.
-   *
-   * @param keyringAccounts - The accounts available in the keyring.
-   * @param scope - Request's scope (CAIP-2).
-   * @param request - Signing request object.
-   * @returns A Promise that resolves to the account address that must
-   * be used to process this signing request, or null if none candidates
-   * could be found.
-   * @throws If the request is invalid.
-   */
-  resolveAccountAddress(
-    keyringAccounts: KeyringAccount[],
-    scope: CaipChainId,
-    request: Infer<typeof BtcWalletRequestStruct>,
-  ): CaipAccountId {
-    const accountsWithThisScope = keyringAccounts.filter((account) =>
-      account.scopes.includes(scope),
-    );
-
-    if (accountsWithThisScope.length === 0) {
-      throw new Error('No accounts with this scope');
-    }
-
-    let addressToValidate: string;
-
-    switch (request.method) {
-      case BtcMethod.BroadcastPsbt:
-      case BtcMethod.FillPsbt:
-      case BtcMethod.ComputeFee:
-      case BtcMethod.GetUtxo:
-      case BtcMethod.SendTransfer:
-      case BtcMethod.SignMessage:
-      case BtcMethod.SignPsbt: {
-        const { account } = request.params;
-        addressToValidate = account.address;
-        break;
-      }
-      default: {
-        throw new Error('Unsupported method');
-      }
-    }
-
-    const foundAccount = accountsWithThisScope.find(
-      (account) => account.address === addressToValidate,
-    );
-
-    if (!foundAccount) {
-      throw new Error('Account not found');
-    }
-
-    return `${scope}:${addressToValidate}`;
   }
 }
